Update the sync status api endpoint
We never passed in oldest_cron to be rendered in the sync status.
I would suggest for clearity that the endpoint returns the next_cron
instead of oldest cron.
since what we are really returning is the number of seconds till the
next cron. This number ideally should never be larger then 60 seconds.

Since we have 2 crons schedules to happen every 1 minute.

If the number is negative we know that the cron is behind and we can
call try to trigger it.